+++ service/history/workflowResetor_test.go
@@ -25,11 +25,9 @@ import (
 	"fmt"
 	"math"
 	"testing"
-	"time"
 
 	"github.com/golang/mock/gomock"
 	"github.com/google/uuid"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
@@ -4072,122 +4070,4 @@ func (s *resetorSuite) TestApplyReset() {
 	s.Empty(resetReq.NewWorkflowSnapshot.SignalInfos)
 	s.Empty(resetReq.NewWorkflowSnapshot.SignalRequestedIDs)
 	s.Equal(0, len(resetReq.NewWorkflowSnapshot.RequestCancelInfos))
-}
-
-func TestFindAutoResetPoint(t *testing.T) {
-	timeSource := clock.NewRealTimeSource()
-
-	// case 1: nil
-	_, pt := execution.FindAutoResetPoint(timeSource, nil, nil)
-	assert.Nil(t, pt)
-
-	// case 2: empty
-	_, pt = execution.FindAutoResetPoint(timeSource, &workflow.BadBinaries{}, &workflow.ResetPoints{})
-	assert.Nil(t, pt)
-
-	pt0 := &workflow.ResetPointInfo{
-		BinaryChecksum: common.StringPtr("abc"),
-		Resettable:     common.BoolPtr(true),
-	}
-	pt1 := &workflow.ResetPointInfo{
-		BinaryChecksum: common.StringPtr("def"),
-		Resettable:     common.BoolPtr(true),
-	}
-	pt3 := &workflow.ResetPointInfo{
-		BinaryChecksum: common.StringPtr("ghi"),
-		Resettable:     common.BoolPtr(false),
-	}
-
-	expiredNowNano := time.Now().UnixNano() - int64(time.Hour)
-	notExpiredNowNano := time.Now().UnixNano() + int64(time.Hour)
-	pt4 := &workflow.ResetPointInfo{
-		BinaryChecksum:   common.StringPtr("expired"),
-		Resettable:       common.BoolPtr(true),
-		ExpiringTimeNano: common.Int64Ptr(expiredNowNano),
-	}
-
-	pt5 := &workflow.ResetPointInfo{
-		BinaryChecksum:   common.StringPtr("notExpired"),
-		Resettable:       common.BoolPtr(true),
-		ExpiringTimeNano: common.Int64Ptr(notExpiredNowNano),
-	}
-
-	// case 3: two intersection
-	_, pt = execution.FindAutoResetPoint(timeSource, &workflow.BadBinaries{
-		Binaries: map[string]*workflow.BadBinaryInfo{
-			"abc": {},
-			"def": {},
-		},
-	}, &workflow.ResetPoints{
-		Points: []*workflow.ResetPointInfo{
-			pt0, pt1, pt3,
-		},
-	})
-	assert.Equal(t, pt.String(), pt0.String())
-
-	// case 4: one intersection
-	_, pt = execution.FindAutoResetPoint(timeSource, &workflow.BadBinaries{
-		Binaries: map[string]*workflow.BadBinaryInfo{
-			"none":    {},
-			"def":     {},
-			"expired": {},
-		},
-	}, &workflow.ResetPoints{
-		Points: []*workflow.ResetPointInfo{
-			pt4, pt0, pt1, pt3,
-		},
-	})
-	assert.Equal(t, pt.String(), pt1.String())
-
-	// case 4: no intersection
-	_, pt = execution.FindAutoResetPoint(timeSource, &workflow.BadBinaries{
-		Binaries: map[string]*workflow.BadBinaryInfo{
-			"none1": {},
-			"none2": {},
-		},
-	}, &workflow.ResetPoints{
-		Points: []*workflow.ResetPointInfo{
-			pt0, pt1, pt3,
-		},
-	})
-	assert.Nil(t, pt)
-
-	// case 5: not resettable
-	_, pt = execution.FindAutoResetPoint(timeSource, &workflow.BadBinaries{
-		Binaries: map[string]*workflow.BadBinaryInfo{
-			"none1": {},
-			"ghi":   {},
-		},
-	}, &workflow.ResetPoints{
-		Points: []*workflow.ResetPointInfo{
-			pt0, pt1, pt3,
-		},
-	})
-	assert.Nil(t, pt)
-
-	// case 6: one intersection of expired
-	_, pt = execution.FindAutoResetPoint(timeSource, &workflow.BadBinaries{
-		Binaries: map[string]*workflow.BadBinaryInfo{
-			"none":    {},
-			"expired": {},
-		},
-	}, &workflow.ResetPoints{
-		Points: []*workflow.ResetPointInfo{
-			pt0, pt1, pt3, pt4, pt5,
-		},
-	})
-	assert.Nil(t, pt)
-
-	// case 7: one intersection of not expired
-	_, pt = execution.FindAutoResetPoint(timeSource, &workflow.BadBinaries{
-		Binaries: map[string]*workflow.BadBinaryInfo{
-			"none":       {},
-			"notExpired": {},
-		},
-	}, &workflow.ResetPoints{
-		Points: []*workflow.ResetPointInfo{
-			pt0, pt1, pt3, pt4, pt5,
-		},
-	})
-	assert.Equal(t, pt.String(), pt5.String())
 }
